Conflicts:
	ui/build.gradle

+++ bufr/src/main/java/ucar/nc2/iosp/bufr/BufrNumbers.java
@@ -296,7 +296,7 @@ final public class BufrNumbers {
     int g = raf.read();
     int h = raf.read();
 
-    return (1 - ((a & 128) >> 6)) * ((a & 127) << 56 | b << 48 | c << 40 | d << 32 | e << 24 | f << 16 | g << 8 | h);
+    return (1 - ((a & 128) >> 6)) * ((long)(a & 127) << 56 | (long) b << 48 | (long) c << 40 | (long) d << 32 | e << 24 | f << 16 | g << 8 | h);
 
    }
 }
